+++ common/pam.c
@@ -64,7 +64,7 @@ pam_t *pam_create_from_file(const char *inpath)
             continue;
 
         size_t linelen = strlen(line);
-        for (int idx = 0; idx < linelen; idx++) {
+        for (int idx = 0; idx < (int)linelen; idx++) {
             if (line[idx] == ' ') {
                 line[idx] = 0;
                 if (tok1) {
@@ -141,7 +141,7 @@ pam_t *pam_create_from_file(const char *inpath)
 
     pam->datalen = pam->width * pam->height * pam->depth;
     pam->data = malloc(pam->datalen);
-    if (pam->datalen != fread(pam->data, 1, pam->datalen, infile)) {
+    if (pam->datalen != (int)fread(pam->data, 1, pam->datalen, infile)) {
         printf("pam.c: couldn't read body\n");
         goto fail;
     }
@@ -182,7 +182,7 @@ int pam_write_file(pam_t *pam, const char *outpath)
     fprintf(f, "P7\nWIDTH %d\nHEIGHT %d\nDEPTH %d\nMAXVAL %d\nTUPLTYPE %s\nENDHDR\n",
             pam->width, pam->height, pam->depth, pam->maxval, tupl);
     int len = pam->width * pam->height * pam->depth;
-    if (len != fwrite(pam->data, 1, len, f)) {
+    if (len != (int)fwrite(pam->data, 1, len, f)) {
         fclose(f);
         return -2;
     }
